Fixed inconsistencies and added more documentation about errors that can be returned.
The input points and segments were store without taking into account the
truncating introduced by the scaling factor. That was introducing
approximations big enough to trigger the exception
UnsolvableParabolaEquation  when interpolating the parabola.

The exception UnsolvableParabolaEquation is usually returned because of
approximation. It is kept as a parameter in cases the user decided to
override it. The documentation has been updated to reflect the good
practice regarding this parameter.